+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to CV Manager will be documented in this file.
 
 Format follows [Keep a Changelog](https://keepachangelog.com/), versioning follows [Semantic Versioning](https://semver.org/).
 
+## [1.11.0] - 2026-03-02
+
+### Added
+- Version display in Settings modal footer showing the current running version
+- Language picker button in toolbar for quick language switching
+
+### Changed
+- Removed "Show All" button from toolbar
+- Moved language selection from Settings tab to toolbar dropdown
+
+### Fixed
+- Update banner text was hardcoded in English instead of using the i18n translation system
+
 ## [1.10.6] - 2026-03-02
 
 ### Fixed